Move settings_page.dart to settings/page folder
